backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
google-auth
google-auth-oauthlib
google-cloud-firestore
google-cloud-storage
google-cloud-core
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jws==0.1.3
MarkupSafe==2.1.5
oauth2client
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.20.0
pyjwt
pyparsing==3.1.4
Pyrebase4==4.8.0
python-jwt==2.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
urllib3==1.26.20
uuid==1.30
Werkzeug==3.0.4
